backend/requirements.txt
fastapi
uvicorn
python-dotenv
yt-dlp
requests
python-multipart
# For better async support if needed
aiofiles
